transcription_tool.py/requirements.txt
google-cloud-documentai
Pillow
google-cloud-aiplatform
python-dotenv
vertexai
